pyproject.toml
[tool.poetry]
name = "battleship-rl"
version = "0.1.0"
description = "Advanced Battleship with RL AI and OpenTelemetry"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "battleship", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pygame = "^2.5.0"
pygbag = "^0.8.0"
torch = "2.2.2"
gymnasium = "^0.29.0"
numpy = "^1.24.0"
opentelemetry-api = "^1.21.0"
opentelemetry-sdk = "^1.21.0"
opentelemetry-exporter-otlp = "^1.21.0"
opentelemetry-instrumentation-logging = "^0.42b0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
battleship-cli = "battleship.cli:main"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = ["src"]

[[tool.mypy.overrides]]
module = "battleship.ai.environment"
disable_error_code = ["misc", "type-arg"]

[[tool.mypy.overrides]]
module = "battleship.ai.agent"
disable_error_code = ["misc", "no-untyped-call"]

[[tool.mypy.overrides]]
module = "battleship.telemetry.*"
disable_error_code = ["attr-defined"]

[[tool.mypy.overrides]]
module = "battleship.telemetry.config"
disable_error_code = ["misc"]

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/battleship --cov-report=html --cov-report=term"
